requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
django-appconf==1.0.5
django-braces==1.15.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-swagger==2.2.0
django-user-accounts==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
factory-boy==3.2.1
Faker==18.4.0
gcloud==0.18.3
geonode-user-accounts==1.0.15
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==2.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
jws==0.1.3
MarkupSafe==2.1.3
myapi==0.0.1
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
protobuf==4.23.0rc2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-jwt==4.0.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.7.1
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.23
uuid==1.30
wrapt==1.15.0
